scripts/requirements.txt
typer
pyyaml
